sys/src/libthread: use uintptr for thread allocation sizes (thanks Geoff Collyer)

+++ sys/src/libthread/threadimpl.h
@@ -187,7 +187,7 @@ void	_threadflagrendez(Thread*);
 Proc*	_threadgetproc(void);
 void	_threadsetproc(Proc*);
 void	_threadinitstack(Thread*, void(*)(void*), void*);
-void*	_threadmalloc(long, int);
+void*	_threadmalloc(uintptr, int);
 void	_threadnote(void*, char*);
 void	_threadready(Thread*);
 void*	_threadrendezvous(void*, void*);
